feat: :sparkles: refactor role module paths for improved organization and consistency
