+++ src/modules/role/application/use-cases/commands/archive-role/archive-role.command.ts
