+++ src/modules/role/application/use-cases/commands/create-role/create-role.use-case.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 
-import { CreateRoleCommand } from '@role/application/commands/use-cases/create-role/create-role.command';
+import { CreateRoleCommand } from '@role/application/use-cases/commands/create-role/create-role.command';
 import { ROLE_REPOSITORY } from '@role/domain/constants/injection-tokens';
 import { IRoleRepositoryContract } from '@role/domain/contracts/role-repository.contract';
 import { CreateRoleDomainService } from '@role/domain/domain-service/create-role.domain-service';
