+++ src/modules/role/application/use-cases/commands/destroy-role/destroy-role.command.ts
