+++ src/modules/role/application/use-cases/queries/get-role/get-role.query.ts
