+++ src/modules/role/application/use-cases/queries/get-role/get-role.use-case.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { QueryHandler, IQueryHandler } from '@nestjs/cqrs';
 
-import { GetRoleQuery } from '@role/application/queries/get-role/get-role.query';
+import { GetRoleQuery } from '@role/application/use-cases/queries/get-role/get-role.query';
 import { GetRoleDomainService } from '@role/domain/domain-service/get-role.domain-service';
 import { RoleModel } from '@role/domain/models/role.model';
