+++ src/modules/role/application/use-cases/queries/list-role/list-role.use-case.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 
 import { Criteria } from '@common/domain/criteria/criteria';
-import { ListRoleQuery } from '@role/application/queries/list-role/list-role.query';
+import { ListRoleQuery } from '@role/application/use-cases/queries/list-role/list-role.query';
 import { ListRoleModel } from '@role/domain/models/role-list.model';
 import { ListRoleDomainService } from '@role/domain/domain-service/list-role.domain-service';
