+++ src/modules/role/infrastructure/controllers/commands/archive-role/archive-role.controller.ts
@@ -3,7 +3,7 @@ import { CommandBus } from '@nestjs/cqrs';
 import { MessagePattern, Payload, RpcException } from '@nestjs/microservices';
 
 import { CMDS_HADES } from '@common/infrastructure/controllers/constants';
-import { ArchiveRoleCommand } from '@role/application/commands/use-cases/archive-role/archive-role.command';
+import { ArchiveRoleCommand } from '@role/application/use-cases/commands/archive-role/archive-role.command';
 import { ArchiveRoleInputDto } from '@role/infrastructure/controllers/commands/archive-role/archive-role-input.dto';
 import { ArchiveRoleOutputDto } from '@role/infrastructure/controllers/commands/archive-role/archive-role-output.dto';
 import { RoleModel } from '@role/domain/models/role.model';
